requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-crispy-forms==1.7.2
django-modelcluster==4.1
django-taggit==0.22.2
django-treebeard==4.2.0
djangorestframework==3.7.7
draftjs-exporter==2.0.0
html5lib==0.999999999
idna==2.6
olefile==0.45.1
Pillow==4.3.0
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
wagtail==2.0
webencodings==0.5.1
Willow==1.1
